requirements.txt
fastapi~=0.111.0
starlette~=0.37.2
pyserial~=3.5
board~=1.0
machine~=0.0.1
smbus==1.1.post2
